requirements.txt
# PetBuddy - Production Dependencies
# Installiere mit: pip install -r requirements.txt

flet==0.28.3
pillow>=10.0.0
python-dotenv>=1.0.0
supabase>=2.0.0

# KI-Rassenerkennung
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0
